+++ scripts/FIRST/SkyNode_FIRST_EXT.sql
@@ -108,7 +108,7 @@ CREATE TABLE dbo.PhotoObj
 	[cz] [float] NOT NULL,
 
 	--/ <summary>HTM ID (J2000)</summary>
-	[htmid] [float] NOT NULL,
+	[htmid] [bigint] NOT NULL,
 
 	--/ <summary> Zone ID </summary>
 	[zoneid] int NOT NULL,
